LayoutTests/js/dom/const.html

<!DOCTYPE HTML>
<html>
<head>
<title>Const test</title>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="bodyId">
<script>
description(
"This test checks that const declarations in JavaScript work and are readonly."
);

// Make sure we don't override properties placed on the global object
var ranConstInitialiser = false;
const bodyId = (ranConstInitialiser = true, "Const initialiser overwrote existing property");
shouldBe("bodyId", "document.getElementById('bodyId')");
shouldBeTrue("ranConstInitialiser");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
