LayoutTests/js/dom/const-expected.txt

This test checks that const declarations in JavaScript work and are readonly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bodyId is document.getElementById('bodyId')
PASS ranConstInitialiser is true
PASS successfullyParsed is true

TEST COMPLETE
